# api/openapi.yaml
openapi: 3.1.0
info:
  title: ame API
  description: Adaptive Mastery Engine — quiz & exam platform with agent identity.
  license:
    name: ''
  version: 0.1.0
paths:
...
  /v1/assessments:
    get:
      tags:
      - assessments
      operationId: list_assessments
      parameters:
      - name: mode
        in: query
        description: Filter by mode (practice or graded)
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: Filter by status (draft, active, archived). When absent, returns public-active + caller's own.
        required: false
        schema:
          type: string
      - name: limit
        in: query
        description: 'Page size (default: 50)'
        required: false
        schema:
          type: integer
          format: int64
      - name: offset
        in: query
        description: Page offset
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: List of assessments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAssessmentsResponse'
        '401':
          description: Missing or invalid token
      security:
      - bearer_auth: []
    post:
      tags:
      - assessments
      operationId: create_assessment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssessmentRequest'
        required: true
      responses:
        '201':
          description: Assessment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSummary'
        '401':
          description: Missing or invalid token
        '422':
          description: Validation failed
      security:
      - bearer_auth: []
...
  /v1/assessments/{id}:
    get:
      tags:
      - assessments
      operationId: get_assessment
      parameters:
      - name: id
        in: path
        description: Assessment ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentDetail'
        '401':
          description: Missing or invalid token
        '404':
          description: Assessment not found
      security:
      - bearer_auth: []
    delete:
      tags:
      - assessments
      operationId: delete_assessment
      parameters:
      - name: id
        in: path
        description: Assessment ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Assessment deleted successfully
        '401':
          description: Missing or invalid token
        '404':
          description: Assessment not found
      security:
      - bearer_auth: []
    patch:
      tags:
      - assessments
      operationId: patch_assessment
      parameters:
      - name: id
        in: path
        description: Assessment ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssessmentRequest'
        required: true
      responses:
        '200':
          description: Assessment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '401':
          description: Missing or invalid token
        '403':
          description: Not the owner
        '404':
          description: Assessment not found
        '422':
          description: Validation failed
      security:
      - bearer_auth: []
  /v1/assessments/{id}/questions:
    post:
      tags:
      - assessments
      operationId: add_assessment_question
      parameters:
      - name: id
        in: path
        description: Assessment ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAssessmentQuestionBody'
        required: true
      responses:
        '201':
          description: Question linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddAssessmentQuestionResponse'
        '401':
          description: Unauthorized
        '404':
          description: No such assessment or question
        '422':
          description: Validation failed
      security:
      - bearer_auth: []
...
  /v1/sessions:
    get:
      tags:
      - crate::http::sessions
      operationId: list_my_sessions
      parameters:
      - name: assessmentId
        in: query
        description: Filter by assessment
        required: false
        schema:
          type: string
          format: uuid
      - name: quizId
        in: query
        description: Filter by quiz (legacy)
        required: false
        schema:
          type: string
          format: uuid
      - name: examId
        in: query
        description: Filter by exam (legacy)
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Attempt history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMySessionsResponse'
        '401':
          description: Unauthorized
      security:
      - bearer_auth: []
    post:
      tags:
      - crate::http::sessions
      operationId: create_session
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionBody'
        required: true
      responses:
        '201':
          description: Created session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '401':
          description: Missing or invalid token
        '403':
          description: Token lacks required scope
        '422':
          description: Validation failed
      security:
      - bearer_auth: []
...
  /v1/sessions/{id}/finish:
    post:
      tags:
      - crate::http::sessions
      operationId: finish
      parameters:
      - name: id
        in: path
        description: Session id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Finished session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinishSessionResponse'
        '404':
          description: No such session
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    AddAssessmentQuestionBody:
      type: object
      properties:
        kind:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/QuestionKind'
            description: Create a new question inline (kind + prompt required if question_id absent).
        pointsOverride:
          type:
          - integer
          - 'null'
          format: int32
        prompt:
          type:
          - string
          - 'null'
        questionId:
          type:
          - string
          - 'null'
          format: uuid
          description: Link an existing bank question by ID.
    AddAssessmentQuestionResponse:
      type: object
      required:
      - questionId
      - orderIndex
      properties:
        orderIndex:
          type: integer
          format: int32
        questionId:
          type: string
          format: uuid
...
    Assessment:
      type: object
      required:
      - id
      - title
      - mode
      - status
      - objectives
      - totalPoints
      - showResultsDuring
      - affectsRating
      - visibility
      - method
      - createdBy
      - createdAt
      - updatedAt
      properties:
        affectsRating:
          type: boolean
        compositionTrace: {}
        course:
          type:
          - string
          - 'null'
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        description:
          type:
          - string
          - 'null'
        durationMin:
          type:
          - integer
          - 'null'
          format: int32
        id:
          type: string
          format: uuid
        method:
          type: string
        mode:
          $ref: '#/components/schemas/AssessmentMode'
        objectives:
          type: array
          items:
            type: string
        passingPoints:
          type:
          - integer
          - 'null'
          format: int32
        showResultsDuring:
          type: boolean
        status:
          $ref: '#/components/schemas/AssessmentStatus'
        timeLimitSeconds:
          type:
          - integer
          - 'null'
          format: int32
        title:
          type: string
        totalPoints:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
        visibility:
          $ref: '#/components/schemas/AssessmentVisibility'
    AssessmentDetail:
      allOf:
      - $ref: '#/components/schemas/Assessment'
      - type: object
        required:
        - sections
        - questions
        properties:
          questions:
            type: array
            items:
              $ref: '#/components/schemas/AssessmentQuestion'
          sections:
            type: array
            items:
              $ref: '#/components/schemas/AssessmentSectionDetail'
    AssessmentMode:
      type: string
      enum:
      - practice
      - graded
    AssessmentQuestion:
      type: object
      required:
      - id
      - kind
      - prompt
      - payload
      - points
      - status
      - orderIndex
      properties:
        codeSnippet: {}
        explanation:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        kind:
          type: string
        orderIndex:
          type: integer
          format: int32
        payload: {}
        points:
          type: integer
          format: int32
        prompt:
          type: string
        status:
          type: string
    AssessmentSectionDetail:
      type: object
      required:
      - id
      - title
      - orderIndex
      - weight
      - itemsCount
      - questions
      properties:
        id:
          type: string
          format: uuid
        itemsCount:
          type: integer
          format: int32
        mix: {}
        orderIndex:
          type: integer
          format: int32
        questions:
          type: array
          items:
            $ref: '#/components/schemas/AssessmentQuestion'
        title:
          type: string
        weight:
          type: number
          format: double
    AssessmentStatus:
      type: string
      enum:
      - draft
      - active
      - archived
    AssessmentSummary:
      type: object
      required:
      - id
      - title
      - mode
      - status
      - visibility
      - objectives
      - totalPoints
      - questionCount
      - completed
      - createdAt
      - updatedAt
      properties:
        completed:
          type: boolean
        course:
          type:
          - string
          - 'null'
        createdAt:
          type: string
          format: date-time
        description:
          type:
          - string
          - 'null'
        durationMin:
          type:
          - integer
          - 'null'
          format: int32
        id:
          type: string
          format: uuid
        lastSessionId:
          type:
          - string
          - 'null'
          format: uuid
        mode:
          type: string
        objectives:
          type: array
          items:
            type: string
        questionCount:
          type: integer
          format: int64
        status:
          type: string
        title:
          type: string
        totalPoints:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
        visibility:
          type: string
    AssessmentVisibility:
      type: string
      enum:
      - public
      - private
...
    CreateAssessmentRequest:
      type: object
      required:
      - title
      - mode
      - objectives
      - method
      properties:
        affectsRating:
          type: boolean
        course:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        durationMin:
          type:
          - integer
          - 'null'
          format: int32
        method:
          type: string
        mode:
          $ref: '#/components/schemas/AssessmentMode'
        objectives:
          type: array
          items:
            type: string
        passingPoints:
          type:
          - integer
          - 'null'
          format: int32
        showResultsDuring:
          type: boolean
        timeLimitSeconds:
          type:
          - integer
          - 'null'
          format: int32
        title:
          type: string
        visibility:
          $ref: '#/components/schemas/AssessmentVisibility'
...
    CreateSessionBody:
      type: object
      properties:
        assessmentId:
          type:
          - string
          - 'null'
          format: uuid
        cats:
          type: array
          items:
            type: string
        count:
          type:
          - integer
          - 'null'
          minimum: 0
        diff:
          type:
          - string
          - 'null'
        duration:
          type:
          - integer
          - 'null'
          format: int64
        examId:
          type:
          - string
          - 'null'
          format: uuid
        mode:
          type:
          - string
          - 'null'
        quizId:
          type:
          - string
          - 'null'
          format: uuid
        tags:
          type: array
          items:
            type: string
        types:
          type: array
          items:
            $ref: '#/components/schemas/QuestionKind'
    CreateSessionResponse:
      type: object
      required:
      - sessionId
      - questions
      properties:
        questions:
          type: array
          items:
            $ref: '#/components/schemas/SessionQuestion'
        sessionId:
          type: string
          format: uuid
...
    FinishSessionResponse:
      type: object
      required:
      - session
      - result
      properties:
        result:
          $ref: '#/components/schemas/SessionResult'
        session:
          $ref: '#/components/schemas/Session'
...
    ListAssessmentsResponse:
      type: object
      required:
      - assessments
      - total
      properties:
        assessments:
          type: array
          items:
            $ref: '#/components/schemas/AssessmentSummary'
        total:
          type: integer
          format: int64
...
    ListMySessionsResponse:
      type: object
      required:
      - sessions
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionSummary'
...
    PlanItem:
      type: object
      required:
      - question_id
      - version
      properties:
        option_order:
          type:
          - array
          - 'null'
          items:
            type: integer
            minimum: 0
        question_id:
          type: string
          format: uuid
        section:
          type:
          - string
          - 'null'
        version:
          type: integer
          format: int32
...
    QuestionKind:
      type: string
      enum:
      - mc
      - tf
      - short
      - essay
      - code
    QuestionPlan:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlanItem'
...
    Session:
      type: object
      required:
      - id
      - user_id
      - kind
      - question_plan
      - status
      - affects_rating
      - rating_snapshot
      - started_at
      properties:
        affects_rating:
          type: boolean
        assessment_id:
          type:
          - string
          - 'null'
          format: uuid
        course_title:
          type:
          - string
          - 'null'
        deadline_at:
          type:
          - string
          - 'null'
          format: date-time
        exam_id:
          type:
          - string
          - 'null'
          format: uuid
        filter:
          type: object
        finished_at:
          type:
          - string
          - 'null'
          format: date-time
        id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/SessionKind'
        question_plan:
          $ref: '#/components/schemas/QuestionPlan'
        quiz_id:
          type:
          - string
          - 'null'
          format: uuid
        quiz_title:
          type:
          - string
          - 'null'
        rating_snapshot:
          type: object
        result:
          type: object
        started_at:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/SessionStatus'
        user_id:
          type: string
          format: uuid
    SessionKind:
      type: string
      enum:
      - quiz
      - exam
      - practice
    SessionQuestion:
      type: object
      required:
      - id
      - kind
      - prompt
      - version
      - points
      - payload
      properties:
        id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/QuestionKind'
        option_order:
          type:
          - array
          - 'null'
          items:
            type: integer
            minimum: 0
        payload:
          type: object
        points:
          type: integer
          format: int32
        prompt:
          type: string
        version:
          type: integer
          format: int32
    SessionResult:
      type: object
      required:
      - points_awarded
      - max_points
      - percent
      properties:
        max_points:
          type: integer
          format: int32
        percent:
          type: number
          format: double
        points_awarded:
          type: integer
          format: int32
    SessionStatus:
      type: string
      enum:
      - in_progress
      - finished
      - abandoned
    SessionSummary:
      type: object
      required:
      - id
      - kind
      - status
      - startedAt
      - attemptNumber
      - totalAttempts
      properties:
        assessmentId:
          type:
          - string
          - 'null'
          format: uuid
        attemptNumber:
          type: integer
          format: int64
        examId:
          type:
          - string
          - 'null'
          format: uuid
        finishedAt:
          type:
          - string
          - 'null'
          format: date-time
        id:
          type: string
          format: uuid
        kind:
          type: string
        maxPoints:
          type:
          - number
          - 'null'
          format: double
        pointsAwarded:
          type:
          - number
          - 'null'
          format: double
        quizId:
          type:
          - string
          - 'null'
          format: uuid
        quizTitle:
          type:
          - string
          - 'null'
        startedAt:
          type: string
          format: date-time
        status:
          type: string
        totalAttempts:
          type: integer
          format: int64
...
    UpdateAssessmentRequest:
      type: object
      properties:
        description:
          type:
          - string
          - 'null'
        objectives:
          type:
          - array
          - 'null'
          items:
            type: string
        status:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/AssessmentStatus'
        title:
          type:
          - string
          - 'null'
        visibility:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/AssessmentVisibility'
...